refactor: move assertion service to separate module
