+++ chk/modules/validate/__init__.py
@@ -3,10 +3,8 @@ Validate module
 """
 
 import dataclasses
-import datetime
 import enum
-import typing
-from collections import abc, UserDict
+from collections import abc
 
 from chk.infrastructure.document import VersionedDocument, VersionedDocumentSupport
 from chk.infrastructure.file_loader import FileContext, ExecuteContext
@@ -20,6 +18,7 @@ from chk.infrastructure.symbol_table import (
     VariableTableManager,
     replace_value,
 )
+from chk.modules.validate.assertion_services import AssertionEntry
 
 VERSION_SCOPE = ["validation"]
 
@@ -49,53 +48,6 @@ ASSERTS_SCHEMA = {
         "valuesrules": {"type": "dict"},
     }
 }
-
-
-class SingleTestRunResult(UserDict):
-    """Result of an assertion run"""
-
-    is_pass: bool
-    time_start: datetime.datetime
-    time_end: datetime.datetime
-    message: str
-
-
-class AllTestRunResult(UserDict):
-    """Result of a test run"""
-
-    id: str
-    time_start: datetime.datetime
-    time_end: datetime.datetime
-    count_all: int
-    results: list[SingleTestRunResult]
-    count_fail: int = 0
-
-    @property
-    def is_all_pass(self) -> bool:
-        """Have all assertion passed for this test run"""
-
-        return self.count_fail == 0
-
-
-class AssertionEntry(typing.NamedTuple):
-    """AssertionEntry holds one assertion operation"""
-
-    assert_type: str
-    type_of_actual: object
-    actual: typing.Any
-    type_of_expected: object
-    expected: typing.Any
-    msg_pass: str
-    msg_fail: str
-    assert_id: str = ""
-
-
-class AssertionAntiquary:
-    """AssertionAntiquary is service class that run assertion"""
-
-    @staticmethod
-    def test_run():
-        pass
 
 
 @dataclasses.dataclass(slots=True)
